feat(postgres): Complete basic setup of the 4 basic GQL queries
