--- src/resolvers/field/index.js
@@ -1,7 +1,0 @@
-const { node } = require('./node')
-
-const fieldResolvers = {
-  node: (_, args) => node(_, args)
-}
-
-module.exports = fieldResolvers
